move standards to standard directory
