+++ framework/packages/abstract-app/src/msgs.rs
@@ -29,22 +29,17 @@
 /// ```
 macro_rules! app_msg_types {
     ($app_type:ty, $app_execute_msg: ty, $app_query_msg: ty) => {
-        /// Abstract App instantiate msg
+        /// Top-level Abstract App instantiate message. This is the message that is passed to the `instantiate` entrypoint of the smart-contract.
         pub type InstantiateMsg =
             <$app_type as ::abstract_sdk::base::InstantiateEndpoint>::InstantiateMsg;
+        /// Top-level Abstract App execute message. This is the message that is passed to the `execute` entrypoint of the smart-contract.
         pub type ExecuteMsg = <$app_type as ::abstract_sdk::base::ExecuteEndpoint>::ExecuteMsg;
+        /// Top-level Abstract App query message. This is the message that is passed to the `query` entrypoint of the smart-contract.
         pub type QueryMsg = <$app_type as ::abstract_sdk::base::QueryEndpoint>::QueryMsg;
+        /// Top-level Abstract App migrate message. This is the message that is passed to the `migrate` entrypoint of the smart-contract.
         pub type MigrateMsg = <$app_type as ::abstract_sdk::base::MigrateEndpoint>::MigrateMsg;
 
         impl ::abstract_core::app::AppExecuteMsg for $app_execute_msg {}
         impl ::abstract_core::app::AppQueryMsg for $app_query_msg {}
-    };
-}
-
-#[deprecated(since = "0.16.2", note = "use app_msg_types instead")]
-#[macro_export]
-macro_rules! app_messages {
-    ($app_type:ty, $app_execute_msg: ty, $app_query_msg: ty) => {
-        $crate::app_msg_types!($app_type, $app_execute_msg, $app_query_msg);
     };
 }
