+++ framework/packages/standards/dex/.editorconfig
