+++ framework/packages/standards/staking/.editorconfig
