+++ framework/packages/standards/staking/src/error.rs
