+++ modules/contracts/adapters/dex/src/lib.rs
@@ -4,13 +4,12 @@ pub mod contract;
 mod exchanges;
 pub(crate) mod handlers;
 pub mod state;
+pub mod msg;
 
 // Export interface for use in SDK modules
 pub use crate::api::DexInterface;
 //:{Dex, DexInterface};
 pub const EXCHANGE: &str = "abstract:dex";
-
-pub use abstract_dex_adapter_traits::msg;
 
 #[cfg(any(feature = "juno", feature = "osmosis"))]
 pub mod host_exchange {
