+++ modules/contracts/apps/dca/src/msg.rs
@@ -1,3 +1,5 @@
+#![warn(missing_docs)]
+
 use abstract_core::objects::{AssetEntry, DexName, PoolReference};
 use abstract_dex_adapter::msg::OfferAsset;
 use cosmwasm_schema::QueryResponses;
@@ -6,16 +8,6 @@ use croncat_app::croncat_integration_utils::CronCatInterval;
 
 use crate::{contract::DCAApp, state::DCAEntry};
 
-// This is used for type safety
-// The second part is used to indicate the messages are used as the apps messages
-// This is equivalent to
-// pub type InstantiateMsg = <App as abstract_sdk::base::InstantiateEndpoint>::InstantiateMsg;
-// pub type ExecuteMsg = <App as abstract_sdk::base::ExecuteEndpoint>::ExecuteMsg;
-// pub type QueryMsg = <App as abstract_sdk::base::QueryEndpoint>::QueryMsg;
-// pub type MigrateMsg = <App as abstract_sdk::base::MigrateEndpoint>::MigrateMsg;
-
-// impl app::AppExecuteMsg for AppExecuteMsg {}
-// impl app::AppQueryMsg for AppQueryMsg {}
 abstract_app::app_msg_types!(DCAApp, DCAExecuteMsg, DCAQueryMsg);
 
 #[cosmwasm_schema::cw_serde]
